v2.24 - CRON Fixed
Finally found what was bugging the cron. Seems it was loading in the statics.php and that was giving issues for some reason. It should now be fixed.

Also added more info to the analytics page.
It now has the last few days and you can switch to graph mode!!!

Analytics (Admin):
 - Added graph mode
 - Added visitors last few days in overview mode

Dashboard (Admin):
 - Fixed a few minor issues with things not loading properly

Config:
 - Updated the config example to now reflect the updated configs